Store numbers in getNumbers function
initial an array that will hold the valid inputs and set the array length to 2 so that only two numbers are passed. While the first element of the array is empty ask for first number. if the inputted value is not valid an invalid message will be printed and be prompted to enter the first number again. Once the first value is valid the for loop is iterated to next slot in the array and the same methodology is used. when both elements have been verified to be numbers the array is returned